+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Discriminated Model union
+- Discriminated Model union #2
 
 ## [0.1.2] - 2024-05-16
 ### Added
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fields auto-generated docstrings
 
 ## Changed
-- Allow Python>=3.10,<3.13
+- Allow Python>=3.10,<3.13 #1
 
 ## [0.1.0] - 2024-05-13
 ### Added